Fix get_visual argument order bug
Ensure that the `start_row`, `start_col`, `end_row`, and `end_col` arguments to `nvim_buf_get_text`
are passed in the correct order regardless of the direction of the visual selection.  Previously,
if the end of the visual selection was before the start, `nvim_buf_get_text` would raise an error
saying "start_col must be less than end_col".